pyproject.toml
[tool.poetry]
name = "competitor-bot-standalone"
version = "0.1.0"
description = ""
authors = ["Alex"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.12"
openai = "^1.82.0"
google-api-python-client = "^2.93.0"
fastapi = "^0.95.2"
pydantic = "^1.10.7"
requests = "^2.31.0"
apscheduler = "^3.10.4"
slack-bolt = "^1.23.0"
boto3 = "^1.38.46"
google-auth = "^2.40.3"
pymupdf = "^1.26.1"

[tool.poetry.scripts]
main2 = "app.main2:main"

[tool.poetry.group.dev.dependencies]
uvicorn = {extras = ["standard"], version = "^0.34.3"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
